refactor: extract shared useChapters hook to eliminate DRY violation (#155, #195)
- Create useChapters() hook in lib/editor-page/ consolidating duplicated
  chapter detection logic (DOM-first with Markdown fallback, 10s auto-refresh)
- Refactor Outline.tsx to use useChapters (removed ~20 lines)
- Refactor ChaptersPanel.tsx to use useChapters (removed ~15 lines)

Closes #155
Closes #195
