+++ lib/editor-page/index.ts
@@ -2,6 +2,7 @@ export { useTextStatistics } from "./use-text-statistics";
 export { useEditorSettings } from "./use-editor-settings";
 export { useElectronEvents } from "./use-electron-events";
 export { useProjectLifecycle } from "./use-project-lifecycle";
+export { useChapters } from "./use-chapters";
 
 export type { TextStatisticsResult } from "./use-text-statistics";
 export type { EditorSettings, EditorSettingsHandlers, EditorSettingsSetters, UseEditorSettingsResult } from "./use-editor-settings";
